We can compile again, and render models.
