+++ .gitignore
@@ -69,3 +69,9 @@ hs_err_pid*
 /site/app/f4769f9bdb7466be65088239c12046d1.eot
 /site/yarn.lock
 /site/app/scripts/bundle.js
+/site/app/674f50d287a8c48dc19ba404d20fe713.eot
+/site/app/912ec66d7572ff821749319396470bde.svg
+/site/app/af7ae505a9eed503f8b8e6982036873e.woff2
+/site/app/fee66e712a8a08eef5805a46892932ad.woff
+/site/app/b06871f281fee6b241d60582ae9369b9.ttf
+/site/app/fontawesome-webfont.svg
