--- node.js/bower.json
@@ -1,24 +1,0 @@
-{
-  "name": "polymer-starter-kit",
-  "private": true,
-  "dependencies": {
-    "ace-builds": "^1.2.0",
-    "file-saver.js": "^1.20150507.2",
-    "iron-elements": "PolymerElements/iron-elements#^1.0.0",
-    "neon-elements": "PolymerElements/neon-elements#^1.0.0",
-    "paper-elements": "PolymerElements/paper-elements#^1.0.1",
-    "pegjs": "^0.9.0",
-    "platinum-elements": "PolymerElements/platinum-elements#^1.0.1",
-    "javascript-detect-element-resize": "sdecima/javascript-detect-element-resize#~0.5.3",
-    "paper-badge": "PolymerElements/paper-badge#^1.1.4",
-    "jquery": "^3.1.1",
-    "paper-tooltip": "PolymerElements/paper-tooltip#^1.1.3",
-    "cytoscape": "^2.7.13",
-    "ansi_up": "^1.3.0",
-    "cytoscape-panzoom": "^2.4.0",
-    "lodash": "^4.17.4"
-  },
-  "resolutions": {
-    "jquery": "^3.1.1"
-  }
-}
