--- site/app/elements/_template-element/_template-element.html
@@ -1,36 +1,0 @@
-<link rel="import" href="../../../bower_components/polymer/polymer.html">
-
-<dom-module id="_template-element">
-  <style>
-    :host {
-      display: block;
-    }
-  </style>
-  <template>
-
-  </template>
-</dom-module>
-<script>
-  (function() {
-    Polymer({
-      is: '_template-element',
-
-      properties: {
-        /**
-         * JS Doc
-         */
-         field: {
-           type: String,
-           value: ''
-         }
-      },
-
-      /**
-       * JS Doc
-       */
-      ready: function(){
-
-      }
-    });
-  })();
-</script>
