--- site/app/elements/_template-element/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html>
-<head>
-
-  <title>_template-element</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-  <script src="../../../bower_components/webcomponentsjs/webcomponents-lite.js"></script>
-  <link rel="import" href="../../../bower_components/iron-component-page/iron-component-page.html">
-
-  <link rel="import" href="_template-element.html">
-
-</head>
-<body>
-
-  <_template-element></_template-element>
-
-</body>
-</html>
