--- site/app/elements/_template-element/test/_template-element-test.html
@@ -1,47 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <title>_template-element</title>
-
-  <script src="../../../../bower_components/webcomponentsjs/webcomponents.js"></script>
-  <script src="../../../../bower_components/web-component-tester/browser.js"></script>
-  <script src="../../../../bower_components/test-fixture/test-fixture-mocha.js"></script>
-
-  <link rel="import" href="../../../../bower_components/polymer/polymer.html">
-  <link rel="import" href="../../../../bower_components/promise-polyfill/promise-polyfill-lite.html">
-  <link rel="import" href="../../../../bower_components/test-fixture/test-fixture.html">
-
-  <link rel="import" href="../_template-element.html">
-</head>
-<body>
-  <test-fixture id="test1">
-    <template>
-      <_template-element></_template-element>
-    </template>
-  </test-fixture>
-
-  <script>
-    suite('<_template-element>', function () {
-
-      setup(function() {
-
-      });
-
-      teardown(function() {
-
-      });
-
-      suite('Test Group Name', function() {
-        test('test 1 name', function() {
-          // do test
-        });
-
-        test('test 2 name', function() {
-          // do test
-        });
-      });
-    });
-  </script>
-
-</body>
-</html>
