--- site/app/elements/_template-element/test/index.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <script src="../../../../bower_components/webcomponentsjs/webcomponents.js"></script>
-    <script src="../../../../bower_components/web-component-tester/browser.js"></script>
-  </head>
-  <body>
-    <script>
-      WCT.loadSuites([
-        '_template-element-test.html'
-      ]);
-    </script>
-  </body>
-</html>
