--- site/app/elements/elements.html
@@ -1,49 +1,0 @@
-<script src="../../bower_components/ace-builds/src/ace.js"></script>
-<script src="../../bower_components/ace-builds/src/ext-language_tools.js"></script>
-<!-- Iron elements -->
-<link rel="import" href="../../bower_components/iron-flex-layout/classes/iron-flex-layout.html">
-<link rel="import" href="../../bower_components/iron-media-query/iron-media-query.html">
-<link rel="import" href="../../bower_components/iron-icons/iron-icons.html">
-<link rel="import" href="../../bower_components/iron-icons/av-icons.html">
-
-<!-- Paper elements -->
-<link rel="import" href="../../bower_components/paper-styles/paper-styles-classes.html">
-<link rel="import" href="../../bower_components/paper-toolbar/paper-toolbar.html">
-<link rel="import" href="../../bower_components/paper-button/paper-button.html">
-<link rel="import" href="../../bower_components/paper-icon-button/paper-icon-button.html">
-<link rel="import" href="../../bower_components/paper-dialog/paper-dialog.html">
-<link rel="import" href="../../bower_components/paper-dialog-scrollable/paper-dialog-scrollable.html">
-<link rel="import" href="../../bower_components/paper-tabs/paper-tabs.html">
-<link rel="import" href="../../bower_components/iron-pages/iron-pages.html">
-<link rel="import" href="../../bower_components/paper-tabs/paper-tab.html">
-<link rel="import" href="../../bower_components/paper-badge/paper-badge.html">
-<link rel="import" href="../../bower_components/paper-checkbox/paper-checkbox.html">
-
-
-<!-- Neon elements -->
-<link rel="import" href="../../bower_components/neon-animation/neon-animated-pages.html">
-
-<!-- Platinum elements -->
-<!-- <link rel="import" href="../bower_components/platinum-sw/platinum-sw-cache.html"> -->
-<!-- <link rel="import" href="../bower_components/platinum-sw/platinum-sw-register.html"> -->
-
-<!-- CSS and JS Imports -->
-<link rel="import" href="../styles/app-theme.html">
-<link rel="import" href="../styles/shared-styles.html">
-<link rel="import" href="imports/file-saver.html">
-
-<!-- Custom Elements -->
-<link rel="import" href="automata-parser/automata-parser.html">
-<link rel="import" href="process-visualisation/process-visualisation.html">
-<link rel="import" href="text-editor/text-editor.html">
-<link rel="import" href="console-logger/console-logger.html">
-<link rel="import" href="automata-walker/automata-walker.html">
-<link rel="import" href="process-selector/process-selector.html">
-<link rel="import" href="save-dialog/save-dialog.html">
-<link rel="import" href="settings-dialog/settings-dialog.html">
-<link rel="import" href="help-dialog/help-dialog.html">
-<link rel="import" href="process-modify/process-modify.html">
-
-
-<!-- Configure your routes here -->
-<!-- <link rel="import" href="routing.html"> -->
