--- site/app/elements/process-visualisation/process-visualisation.html
@@ -1,23 +1,0 @@
-<link rel="import" href="../../../bower_components/polymer/polymer.html">
-<link rel="import" href="../../../bower_components/iron-flex-layout/iron-flex-layout.html">
-<link rel="import" href="../imports/element-resizer.html">
-<link rel="stylesheet" href="../../../bower_components/cytoscape-panzoom/cytoscape.js-panzoom.css">
-<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/qtip2/3.0.3/jquery.qtip.min.css">
-<script src="../../../bower_components/lodash/lodash.js"></script>
-<script src="../../../bower_components/jquery/dist/jquery.min.js"></script>
-<script src="graph-constructor.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/cytoscape/2.7.15/cytoscape.js"></script>
-<script src="https://cdn.rawgit.com/cytoscape/cytoscape.js-cose-bilkent/1.0.5/cytoscape-cose-bilkent.js"></script>
-<script src="https://cdn.jsdelivr.net/qtip2/3.0.3/jquery.qtip.min.js"></script>
-<script src="https://cdn.rawgit.com/cytoscape/cytoscape.js-qtip/13da54c9/cytoscape-qtip.js"></script>
-<script src="../../../bower_components/cytoscape-panzoom/cytoscape-panzoom.js"></script>
-<script src="cxt-menu.js"></script>
-
-<dom-module id="process-visualisation">
-  <template>
-    <div id="svg-parent"></div>
-    <div id="name">[[name]]</div>
-  </template>
-</dom-module>
-<script src="script.js"></script>
