--- site/app/elements/save-dialog/save-dialog.html
@@ -1,48 +1,0 @@
-<link rel="import" href="../../../bower_components/polymer/polymer.html">
-<link rel="import" href="../../../bower_components/iron-flex-layout/iron-flex-layout.html">
-<link rel="import" href="../imports/element-resizer.html">
-
-<dom-module id="save-dialog">
-  <template>
-    <paper-dialog id="save-dialog" with-backdrop>
-      <paper-input id="filename" is="iron-input" label="Insert Filename Here" placeholder="untitled"
-                   bind-value="{{inputValue}}" class="middle" title="The name given to the file when it is downloaded."></paper-input>
-      <paper-checkbox checked="{{saveSettings.saveLayout}}">
-        <span>Save layout</span>
-      </paper-checkbox>
-      <paper-checkbox checked="{{saveSettings.saveCode}}">
-        <span>Save code</span>
-      </paper-checkbox>
-      <br/>
-      <paper-button on-tap="save">Save</paper-button>
-      <paper-button style="float:right" on-tap="close">Cancel</paper-button>
-    </paper-dialog>
-  </template>
-</dom-module>
-<script>
-  (function() {
-    'use strict';
-    Polymer({
-      is: 'save-dialog',
-      open: function() {
-        this.saveSettings = app.saveSettings;
-        $('#save-dialog')[0].open();
-      },
-      close: function() {
-        app.saveSettings = this.saveSettings;
-        $('#save-dialog')[0].close();
-      },
-      save: function() {
-        app.saveSettings = this.saveSettings;
-        app.downloadFile();
-        $('#save-dialog')[0].close();
-      },
-      getFileName: function() {
-        return $('#filename')[0].inputElement.bindValue;
-      },
-      opened: function() {
-        return $('#save-dialog')[0].opened;
-      }
-    })
-  })();
-</script>
