--- site/app/elements/text-editor/text-editor.html
@@ -1,24 +1,0 @@
-<link rel="import" href="../../../bower_components/polymer/polymer.html">
-<script src="lexer-tokens.js"></script>
-<script src="syntaxHighligher.js"></script>
-<script src="theme.js"></script>
-<dom-module id="text-editor">
-  <style>
-    :host {
-      display: block;
-      position: relative;
-    }
-    #editor{
-      box-sizing: border-box;
-      position: absolute;
-      top: 0;
-      right: 0;
-      bottom: 0;
-      left: 0;
-    }
-  </style>
-  <template>
-    <div id="editor"></div>
-  </template>
-</dom-module>
-<script src="script.js"></script>
