--- site/app/manifest.json
@@ -1,29 +1,0 @@
-{
-  "name": "Model Checker",
-  "short_name": "Model Checker",
-  "icons": [{
-        "src": "images/touch/icon-128x128.png",
-        "sizes": "128x128",
-        "type": "image/png"
-      }, {
-        "src": "images/touch/apple-touch-icon.png",
-        "sizes": "152x152",
-        "type": "image/png"
-      }, {
-        "src": "images/touch/ms-touch-icon-144x144-precomposed.png",
-        "sizes": "144x144",
-        "type": "image/png"
-      }, {
-        "src": "images/touch/chrome-touch-icon-192x192.png",
-        "sizes": "192x192",
-        "type": "image/png"
-      },{
-        "src": "images/touch/chrome-splashscreen-icon-384x384.png",
-        "sizes": "384x384",
-        "type": "image/png"
-      }],
-  "start_url": "/?homescreen=1",
-  "background_color": "#3E4EB8",
-  "display": "standalone",
-  "theme_color": "#2E3AA1"
-}
