+++ site/app/package.json
@@ -11,10 +11,14 @@
     "bootstrap-webpack": "^0.0.6",
     "css-loader": "^0.28.4",
     "cytoscape": "^3.1.3",
+    "cytoscape-cose-bilkent": "^1.6.13",
     "cytoscape-cxtmenu": "^2.10.3",
+    "cytoscape-panzoom": "^2.5.2",
+    "cytoscape-qtip": "^2.7.1",
     "exports-loader": "^0.6.4",
     "extract-text-webpack-plugin": "^2.1.2",
     "file-loader": "^0.11.2",
+    "font-awesome": "^4.7.0",
     "fs": "^0.0.1-security",
     "imports-loader": "^0.7.1",
     "jquery": "^3.2.1",
