--- site/app/robots.txt
@@ -1,4 +1,0 @@
-# www.robotstxt.org
-
-User-agent: *
-Disallow:
