+++ site/app/scripts/constants.js
@@ -9,10 +9,3 @@ const PROCESS_TYPES = ['automata', 'petrinet'];
 const TAU = '\u03C4';
 const DELTA = '\u03B4';
 const GAMMA = '\u0263';
-const Colours = {
-  red: '#C0392b',
-  blue: '#3498DB',
-  green: '#2ECC71',
-  grey: '#BDC3C7',
-  textBackground: 'rgba(255, 255, 255, 0.5)'
-};
