+++ site/app/scripts/save-load.js
@@ -1,13 +1,30 @@
 module.exports = {
-    load: load
+    init: init,
+    downloadFile: downloadFile
 };
+function loadFile(file) {
+    const reader = new FileReader();
+    reader.readAsText(file);
+    reader.onload = function(file) {
+        app.editor.setCode(file.target.result);
+    }
+}
 /**
  * Listen for key presses.
  * Note: Needs to listen for keydown (not keyup) in order to prevent browser default action
  */
-function load() {
+function init() {
     $("#save-dialog").on('shown.bs.modal', function () {
         $(this).find("#fileName").focus().select();
+    });
+    const files = $('#files');
+    $("#open-bt").click(()=>files.click());
+    files.on('change',function(e) {
+        if (this.value === '') {
+            return;
+        }
+        loadFile(this.files[0]);
+        this.value = '';
     });
     document.addEventListener('keydown', function (e) {
         // if (app.$.help.opened()) {
@@ -51,4 +68,27 @@ function load() {
                 return;
         }
     });
+}
+
+
+/**
+ * Save to code the user has written to their computer (as a download).
+ */
+function downloadFile() {
+    let filename = app.$.save.getFileName();
+    // if filename has not been defined set to untitled
+    if(filename === ''){
+        filename = 'untitled';
+    }
+    let output = "";
+    if (app.settings.getSettings().saveCode)
+        output+= app.$.editor.getCode();
+    if (app.settings.getSettings().saveLayout) {
+        output+="\nvisualiser_json_layout:"
+        output+= JSON.stringify(app.$.visualiser.cy.json());
+    }
+    const blob = new Blob(
+        [output],
+        {type: 'text/plain;charset=utf-8'});
+    saveAs(blob, filename + '.txt');
 }
