--- site/app/test/index.html
@@ -1,27 +1,0 @@
-<!doctype html>
-<html>
-
-  <head>
-    <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-    <title>Elements Test Runner</title>
-    <meta charset="UTF-8">
-
-    <script src="../../bower_components/webcomponentsjs/webcomponents.min.js"></script>
-    <script src="../../bower_components/web-component-tester/browser.js"></script>
-  </head>
-
-  <body>
-
-    <script>
-      WCT.loadSuites([
-        'graph.html',
-        '../elements/automata-parser/test/grammar-test.html',
-        '../elements/automata-parser/test/automata-parser-test.html',
-        '../elements/automata-visualisation/test/automata-visualisation-test.html'
-      ]);
-    </script>
-
-
-  </body>
-
-</html>
